flask_basic/code/requirements.txt
Flask_JWT_Extended
Flask_RESTful
Flask_SQLAlchemy
